Moved templates out of descriptor pool cpp file
