+++ src/VkZero/descriptor_pool.cpp
@@ -1,5 +1,6 @@
 #include "VkZero/Internal/descriptor_pool_internal.hpp"
 #include "VkZero/Internal/core_internal.hpp"
+#include "VkZero/types.hpp"
 #include "shaders.hpp"
 #include <tuple>
 #include <utility>
@@ -13,91 +14,18 @@ using namespace VkZero;
 
 using namespace VkZero;
 
-template <VkDescriptorType Type>
-struct DescriptorTypeObject
-{
-    static constexpr VkDescriptorType value = Type;
-};
+DescriptorPool::DescriptorPool(){}
 
-template <typename Ts>
-struct get_shader_bindings;
-
-// Base case: empty tuple
-template <>
-struct get_shader_bindings<std::tuple<>> {
-    using type = std::tuple<>;
-};
-
-// Recursive case
-template <typename First, typename... Rest>
-struct get_shader_bindings<std::tuple<First, Rest...>> {
-    using type = decltype(
-        std::tuple_cat(
-            std::declval<typename First::BindingsList>(),
-            std::declval<typename get_shader_bindings<std::tuple<Rest...>>::type>()
-        )
-    );
-};
-
-template <typename T>
-struct DescriptorTypes;
-
-template<>
-struct DescriptorTypes<std::tuple<>> {
-    using type = std::tuple<>;
-};
-
-template <typename First, typename... Rest>
-struct DescriptorTypes<std::tuple<First, Rest...>> {
-    using type = decltype(
-        std::tuple_cat(
-            std::tuple<DescriptorTypeObject<(VkDescriptorType)First::type()>>{},
-            typename DescriptorTypes<std::tuple<Rest...>>::type{}
-        )
-    );
-};
-
-template <typename T, typename Tuple>
-struct contains;
-
-template <typename T>
-struct contains<T, std::tuple<>> : std::false_type {};
-
-template <typename T, typename U, typename... Rest>
-struct contains<T, std::tuple<U, Rest...>> 
-    : std::conditional_t<std::is_same_v<T, U>, std::true_type, contains<T, std::tuple<Rest...>>> {};
-
-template <typename Tuple>
-struct unique_types;
-
-template <>
-struct unique_types<std::tuple<>> { using type = std::tuple<>; };
-
-template <typename First, typename... Rest>
-struct unique_types<std::tuple<First, Rest...>> {
-    using Tail = typename unique_types<std::tuple<Rest...>>::type;
-    using type = std::conditional_t<
-        contains<First, Tail>::value,
-        Tail,
-        decltype(std::tuple_cat(std::tuple<First>{}, Tail{}))
-    >;
-};
-
-DescriptorPool::DescriptorPool()
+void DescriptorPool::build(std::vector<DescriptorType> types)
 {
     std::vector<VkDescriptorPoolSize> poolSizes;
-
-
-    using types = typename unique_types<DescriptorTypes<get_shader_bindings<GlobalShaderLibrary::types>::type>::type>::type;
-
-    std::apply([&]<typename... Ts>(Ts...){
-       ([&](VkDescriptorType type){
+    for (auto t : types)
+    {
         VkDescriptorPoolSize poolSize{};
-        poolSize.type = type;
+        poolSize.type = (VkDescriptorType)t;
         poolSize.descriptorCount = MAX_DESCRIPTOR_COUNT;
         poolSizes.push_back(poolSize);
-       }(Ts::value), ...);
-    }, types{});
+    }
 
     VkDescriptorPoolCreateInfo createInfo = {};
     createInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
